requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
config==0.5.1
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==28.4.1
idna==3.8
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
